Functions

Base.maximum — Method

maximum(c::Coefplot)

Compute the maximal value that the Coefplot can reach with its error bar.

Base.maximum — Method

maximum(m::MultiCoefplot)

Compute the maximal value that the MultiCoefplot can reach with its error bar.

Base.minimum — Method

minimum(c::Coefplot)

Compute the minimal value that the Coefplot can reach with its error bar.

Base.minimum — Method

minimum(m::MultiCoefplot)

Compute the minimal value that the MultiCoefplot can reach with its error bar.

Base.parse — Method

Base.parse(r::SupportedEstimation; 
    rename::Vector{Pair{T,R}} where T<:AbstractString where R=missing, 
    keepcoef::Vector{T} where T<:AbstractString=missing, 
    kwargs...)

This function takes the regression result and convert it into a Coefplot. If rename is specified, it will rename the coefficients, and keep them in the order the renameing pairs are specified. If keepcoef is specified, it will filter the coefficients, and keep them in the order they are specified. Named arguements of Coefplot() can also be passed in this function.

Base.sort! — Method

Base.sort!(m::MultiCoefplot)

sort the vector of Coefplots according to csorter.

Coefplots._escape_string — Function

_escape_string(str::AbstractString[, esc]; keep = ())::AbstractString
_escape_string(io, str::AbstractString[, esc]; keep = ())::Nothing

Almost the same as the Base.escape_string except that _escape_string() escapes , , (, ), and . by putting it in a pair of brackets.

Coefplots._parse — Method

_parse(r::SupportedEstimation, ps::Pair{<:AbstractString, <:Any} ...; drop_unmentioned::Bool=true, kwargs...)

This function takes the regression result and convert it into a Coefplot. ps is how you want to rename the coefficients. If drop_unmentioned, parse will drop all the unmentioned coefficient in ps in the Coefplot. Named arguements of Coefplot() can also be passed in this function.

Coefplots._parse — Method

_parse(r::SupportedEstimation, ps::Vector{T} where T<:AbstractString; kwargs...)

This function takes the regression result and convert it into a Coefplot. ps contains what coefficients and in what order the user wants to keep. Named arguements of Coefplot() can also be passed in this function.

Coefplots._parse — Method

_parse(r::SupportedEstimation; kwargs...)

This function takes the regression result and convert it into a Coefplot. Named arguements of Coefplot() can also be passed in this function.

Coefplots.color! — Method

color!(c::Coefplot, clr::Color)

Reset the color of a Coefplot.

Coefplots.errbar_length — Function

errbar_length(data::AbstractDataFrame, level::Real=0.95)

Compute the length of the error bar for each coefficient.

Coefplots.get_axis_options — Method

get_axis_options(c::Coefplot)

Renders the properties of a Coefplot object as options of the \begin{axis}

Coefplots.get_axis_options — Method

get_axis_options(m::MultiCoefplot)

Renders the properties of a MultiCoefplot object as options of the \begin{axis}

Coefplots.get_groupplot_options — Method

get_groupplot_options(g::GroupedCoefplot)

Renders the properties of a GroupedCoefplot object as options of the \begin{groupplot}

Coefplots.get_groupplot_options — Method

get_nextgroupplot_options(g::GroupedMultiCoefplot)

Renders the properties of a GroupedMultiCoefplot object as options of the \begin{groupplot}

Coefplots.get_nextgroupplot_options — Method

get_nextgroupplot_options(c::Coefplot)

Renders the properties of a Coefplot object as options of the \begin{nextgroupplot}

Coefplots.get_nextgroupplot_options — Method

get_nextgroupplot_options(m::MultiCoefplot)

Renders the properties of a MultiCoefplot object as options of the \begin{nextgroupplot}

Coefplots.get_plot_options — Method

get_plot_options(c::Coefplot)

Renders the properties of a Coefplot object as options of the \addplot

Coefplots.latex_escape — Method

latex_escape(s::AbstractString)

Escapes ,, (, ), and . by putting it in a pair of brackets, also esxapes & and % by adding \

Coefplots.sortcoef! — Method

sortcoef!(c::Coefplot; rev::Bool=false)

Sort the coefficient in ascending or descending order. Both c.data and c.sorter is modified.

Coefplots.to_axis — Method

to_axis(c::Coefplot, other::SupportedAddition ...)

Converts the Coefplot object to a PGFPlotsX.Axis object. Other supported components are allowed and appended after the Coefplot within the axis.

Coefplots.to_axis — Method

to_axis(g::GroupedCoefplot, other::SupportedAddition ...)

Converts the GroupedCoefplot object to a PGFPlotsX.Axis object. Other supported components are allowed and appended after the GroupedCoefplot within the axis.

Coefplots.to_axis — Method

to_axis(g::GroupedMultiCoefplot, other::SupportedAddition ...)

Converts the GroupedMultiCoefplot object to a PGFPlotsX.Axis object. Other supported components are allowed and appended after the GroupedMultiCoefplot within the axis.

Coefplots.to_axis — Method

to_axis(m::MultiCoefplot, other::SupportedAddition ...)

Converts the MultiCoefplot object to a PGFPlotsX.Axis object. Other supported components are allowed and appended after the MultiCoefplot within the axis.

Coefplots.to_options — Method

to_options(l::Bar)

generates tikz options related to bars: [draw={}, draw opacity={}, line width=0.6pt}, densely dotted].

Coefplots.to_options — Method

to_options(l::CaptionStyle)

generates PGF options related to caption style: [font={}, rotate={}].

Coefplots.to_options — Method

to_options(l::Label, label::Symbol, label_style::Symbol)

generates PGF options related to label: [label={}, label_style={font={}, rotate={}}].

Coefplots.to_options — Method

to_options(l::Legend)

generates tikz options related to the legend style: [font={}, rotate={}].

Coefplots.to_options — Method

to_options(l::Mark, mark::Union{Symbol, String}, mark_options::Union{Symbol, String})

generates PGF options related to mark: [mark={}, mark_options={fill={}, fill opacity={}, mark size={}, draw={}, draw opacity={}, solid, line width=...}].

Coefplots.to_options — Method

to_options(l::Legend)

generates tikz options related to note: [font={}, rotate={}].

Coefplots.to_picture — Method

to_picture(c::Coefplot, other::SupportedAddition ...)

Convert the Coefplot object to an PGFPlotsX.TikzPicture. Other supported components are allowed and appended after the Coefplot. The note field is drawn as a node beyond the axis.

Coefplots.to_picture — Method

to_picture(g::GroupedCoefplot, other::SupportedAddition ...)

convert the GroupedCoefplot object to an PGFPlotsX.TikzPicture. Other supported components are allowed and appended after the GroupedCoefplot within the axis. Note is added.

Coefplots.to_picture — Method

to_picture(m::GroupedMultiCoefplot, other::SupportedAddition ...)

convert the GroupedMultiCoefplot object to an PGFPlotsX.TikzPicture. Other supported components are allowed and appended after the GroupedMultiCoefplot within the axis. Note is added.

Coefplots.to_picture — Method

to_picture(m::MultiCoefplot, other::SupportedAddition ...)

convert the MultiCoefplot object to an PGFPlotsX.TikzPicture. Other supported components are allowed and appended after the MultiCoefplot within the axis. Note is added.

Coefplots.to_plot — Method

to_plot(c::Coefplot)

Convert the Coefplot object to an PGFPlotsX.AxisElement. It is realized using the PGFPlotsX.Plot/PGFPlotsX.Coordinates combination.