Types

Coefplots.Annotation — Type

Annotation

Constructors

Annotation(;angle, content, point_at)

Keyword arguments

angle::Real is the angle of the pointer of the annotation.

content::String is the textual content of the annotation.

point_at::Tuple{Real, Real} specifies the location in the axis frame that the annotation should point at. A typical one would be (1, 0), which means that the annotation points at the south-east corner of the axis frame.

Coefplots.Bar — Type

Bar

This enters Coefplot.errbar, Coefplot.connect, and that of Multi-, Grouoped-, GroupedMultiCoefplot.

Constructors

Bar(;draw=missing, linewidth=missing, linetype=missing)

Keyword arguments

draw::Union{Color, Missing} is the color of the bar.

linetype::Union{Real, Missing} is the type of line used to draw the bar. Available choices are: "solid", "dotted", "densely dotted", "loosely dotted", "dashed", "densely dashed", "loosely dashed", "dash dot", "densely dash dot", "loosely dash dot", "dash dot dot", "densely dash dot dot", "loosely dash dot dot".

linewidth::Union{Real, Missing} is the width of line used to draw the bar.

Coefplots.CaptionStyle — Type

CaptionStyle

This enters Coefplot.xticklabel, Coefplot.yticklabel, Coefplot.xlabel.captionstyle, Coefplot.ylabel.captionstyle, Coefplot.title.captionstyle, Coefplot.note.captionstyle.

Constructors

CaptionStyle(;font=missing, size=missing, rotate=missing)

Keyword arguments

font::Union{Symbol, String, Missing} contains the font's T1 code

size::Union{Real, Missing} is the font's size in pt.

rotate::Union{Real, Missing} is the rotation angle counterclockwise for the caption.

Coefplots.Coefplot — Type

Coefplot

A Coefplot object. It contains the all information for which a coefplot should be plotted. The keyword arguements of the constructor are all optional.

Constructors

Coefplot(data::AbstractDataFrame; <keyword arguments>)

Arguments

title::Label: the title to the plot.

xlabel::Label: the xlabel to the plot.

ylabel::Label: the ylabel to the plot.

xticklabel::CaptionStyle: the style of the xtick.

yticklabel::CaptionStyle: the style of the ytick.

width::Real = 240: the width of the axis frame

height::Real = 204: the height of the axis frame

keepmark::Bool = true: true if the user wants to plot the point estimates, false otherwise.

keepconnect::Bool = false: true if the user wants to connect the neighboring point estimates, false otherwise.

mark::Mark: the style of mark for the point estimates.

errormark:Mark: the style of mark for the endpoints of confidence interval.

errorbar::Bar: the style of the error bar.

connect::Bar: the style of the line that connects the neighboring point estimates.

offset::Real = 0: similar to that of the Stata package, it shifts the coefplot along the axis that represents coefficient name.

sorter::Vector{String} = String[]: a vector indicating the content and the order of the coefficients. If empty, use the order of the data.varname.

level::Real = 0.95: the confidence level.

note::Union{Note, Missing}: a note that is attached to the south of the plot.

vertical::Bool = true: if true, the errorbars are parallel to y axis; if false, the errorbars are parallel to x axis.

Coefplots.GroupedCoefplot — Type

GroupedCoefplot

A GroupedCoefplot object. It contains the all information for which a grouoped coefplot should be plotted. The keyword arguements of the constructors are all optional.

Constructors

GroupedCoefplot(data::Pair{<:Any, Coefplot}...;  <keyword arguments>)
GroupedCoefplot(gdata::GroupedDataFrame;  <keyword arguments>)

Arguments

title::Label: the title to the plot.

xlabel::Label: the xlabel to the plot.

ylabel::Label: the ylabel to the plot.

xticklabel::CaptionStyle: the style of the xtick.

yticklabel::CaptionStyle: the style of the ytick.

width::Real = 240: the width of the axis frame

height::Real = 204: the height of the axis frame

note::Union{Note, Missing}: a note that is attached to the south of the plot.

vertical::Bool = true: if true, the errorbars are parallel to y axis; if false, the errorbars are parallel to x axis.

Coefplots.GroupedMultiCoefplot — Type

GroupedMultiCoefplot

A GroupedMultiCoefplot object. It contains the all information for which a set of multiple GroupedCoefplots should be plotted together. The keyword arguements of the constructor are all optional.

Constructors

GroupedMultiCoefplot(data::Pair{<:Any, GroupedCoefplot} ...;  <keyword arguments>)
GroupedMultiCoefplot(data::Pair{<:Any, MultiCoefplot} ...;  <keyword arguments>)

Arguments

title::Label: the title to the plot.

xlabel::Label: the xlabel to the plot.

ylabel::Label: the ylabel to the plot.

xticklabel::CaptionStyle: the style of the xtick.

yticklabel::CaptionStyle: the style of the ytick.

show_legend::Union{Vector{Bool}, Missing} = missing: a boolean vector specifying which legend of the subplot should be shown. The default is to show only the first legend.

width::Real = 240: the width of the axis frame

height::Real = 204: the height of the axis frame

interval::Union{Real,Missing} = missing: determines the distance between each Coefplot. Each Coefplot's offset is computed according to this.

note::Union{Note, Missing}: a note that is attached to the south of the plot.

vertical::Bool = true: if true, the errorbars are parallel to y axis; if false, the errorbars are parallel to x axis.

Coefplots.HBand — Method

HBand([options], ymin, ymax)

A horizontal band from ymin to ymax

Coefplots.HLine — Method

HLine([options], y)

A horizontal line at y

Coefplots.Label — Type

Label

This enters Coefplot.xlabel, Coefplot.ylabel, and that of Multi-, Grouoped-, GroupedMultiCoefplot.

Constructors

Label(;content=missing, captionstyle=CaptionStyle())

Keyword arguments

content::Union{String, Missing} is the content of the label.

captionstyle::Union{CaptionStyle, Missing} is the font's style in which the label is printed.

Coefplots.Legend — Type

Legend

This enters MultiCoefplot.legend. It determines the style of the legend. The content of the legend is defined by the title of the Coefplot.

Constructors

Legend(;anchor=missing, at=missing, font=missing, size=missing)

Keyword arguments

anchor::Union{Symbol, String, Missing} specifies the anchor of the legend box that is used for alignment. A typical one would be "north west".

at::Any specifies the location in the axis frame that the anchor should adhere to. A typical one would be (1, 0), which means that the anchor is fixed to the south-east corner of the axis frame.

font::Union{Symbol, String, Missing} is the font in which the legend should be printed in.

size::Union{Real, Missing} is the font size.

Coefplots.Mark — Type

Mark

This enters Coefplot.mark, Coefplot.errormark, and that of Multi-, Grouoped-, GroupedMultiCoefplot.

Constructors

Mark(;mark=missing, marksize=missing, linetype=missing, linewidth=missing, fill=missing, draw=missing)

Keyword arguments

mark::Union{Symbol, String, Missing} is the shape of the mark.

marksize::Union{Real, Missing} is the size of the mark in pt.

linetype::Union{Real, Missing} is the type of line of the outline of the mark. Available choices are: "solid", "dotted", "densely dotted", "loosely dotted", "dashed", "densely dashed", "loosely dashed", "dash dot", "densely dash dot", "loosely dash dot", "dash dot dot", "densely dash dot dot", "loosely dash dot dot".

linewidth::Union{Real, Missing} is the width of line of the outline of the mark.

fill::Union{Color, Missing} is the color used to fill the mark

draw::Union{Color, Missing} is the color used to draw the outline of the mark.

Coefplots.MultiCoefplot — Type

MultiCoefplot

A MultiCoefplot object. It contains the all information for which a set of multiple Coefplots should be plotted together. The keyword arguements of the constructor are all optional.

Constructors

MultiCoefplot(data::Coefplot ...; <keyword arguments>)

Arguments

title::Label: the title to the plot.

xlabel::Label: the xlabel to the plot.

ylabel::Label: the ylabel to the plot.

xticklabel::CaptionStyle: the style of the xtick.

yticklabel::CaptionStyle: the style of the ytick.

legend::Legend: the style of the legend.

width::Real = 240: the width of the axis frame

height::Real = 204: the height of the axis frame

interval::Union{Real,Missing} = missing: determines the distance between each Coefplot. Each Coefplot's offset is computed according to this.

sorter::Vector{String} = String[]: a vector indicating the content and the order of the coefficients. If empty, use the union of the data.varname of each Coefplot.

csorter::Vector{String} = String[]: a vector indicating the order of the coefplots. If empty, use the order of data.

note::Union{Note, Missing}: a note that is attached to the south of the plot.

vertical::Bool = true: if true, the errorbars are parallel to y axis; if false, the errorbars are parallel to x axis.

The following arguements are from Coefplots(), and will be passed down to each Coefplot if specified in MultiCoefplot().

keepmark::Bool = true: true if the user wants to plot the point estimates, false otherwise.

keepconnect::Bool = false: true if the user wants to connect the neighboring point estimates, false otherwise.

mark::Mark: the style of mark for the point estimates.

errormark:Mark: the style of mark for the endpoints of confidence interval.

errorbar::Bar: the style of the error bar.

connect::Bar: the style of the line that connects the neighboring point estimates.

offset::Real = 0: similar to that of the Stata package, it shifts the coefplot along the axis that represents coefficient name.

sorter::Vector{String} = String[]: a vector indicating the content and the order of the coefficients. If empty, use the order of the data.varname.

level::Real = 0.95: the confidence level.

Coefplots.Note — Type

Note <: PGFPlotsX.TikzElement

Constructors

Note(;content=missing, anchor=missing ,at=missing, align=missing, captionstyle=missing)

Keyword arguments

content::Union{String, Missing} is the content of the note.

anchor::Union{Symbol, String, Missing} specifies the anchor of the note box that is used for alignment. A typical one would be "north west".

at::Any specifies the location in the axis frame that the anchor should adhere to. A typical one would be (1, 0), which means that the anchor is fixed to the south-east corner of the axis frame.

align::Union{Symbol, String, Missing} specifies how the note should be aligned. It could be "left", "right" or "center".

captionstyle::Union{captionstyle, Missing} is the caption style of the note.

Coefplots.VBand — Method

VBand([options], xmin, xmax)

A vertical band from xmin to xmax

Coefplots.VLine — Method

VLine([options], x)

A vertical line at x

Coefplots.rHBand — Method

rHBand([options], ymin, ymax)

A horizontal band from ymin to ymax, which are relative and range between 0 and 1.

Coefplots.rHLine — Method

rHLine([options], y)

A horizontal line at y, where y takes a real value between 0 and 1 and denotes the relative position of the horizontal line.

Coefplots.rVBand — Method

rVBand([options], xmin, xmax)

A vertical band from xmin to xmax, which are relative and range between 0 and 1.

Coefplots.rVLine — Method

rVLine([options], x)

A vertical line at x, where x takes a real value between 0 and 1 and denotes the relative position of the vertical line.